Added font + working script
Added a couple extra fonts for the bill. the divs and a's have been switches around so the entire btn/circle is clickable instead of just the text. The product page has been changed slightly. The product amount now multiplies the amount of DB queries it will do for each product. This is a strange way to do it, but since there is no column for "amount" in the db, is this the only option. lastly, a start has been made to the bill page, a bit of styling, and the fetching of the products + prices